# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.MusicResponse:
    properties:
      artist:
        type: string
      bpm:
        type: number
      composer:
        type: string
      description:
        type: string
      genre:
        type: string
      id:
        type: string
      is_free:
        type: boolean
      is_recommended:
        type: boolean
      name:
        type: string
      release_date:
        type: string
      unlock_level:
        type: integer
    type: object
  dto.StageResponse:
    properties:
      difficulty:
        type: integer
      id:
        type: string
      level_name:
        type: string
      max_combo:
        type: integer
      music_id:
        type: string
      total_notes:
        type: integer
    type: object
  entity.CreateRecordRequest:
    properties:
      accuracy:
        type: number
      additional_info:
        additionalProperties: true
        type: object
      bad_count:
        type: integer
      character_id:
        type: string
      good_count:
        type: integer
      is_full_combo:
        type: boolean
      is_perfect_play:
        type: boolean
      max_combo:
        type: integer
      miss_count:
        type: integer
      music_id:
        type: string
      perfect_count:
        type: integer
      play_duration:
        type: integer
      rank:
        type: string
      score:
        minimum: 0
        type: integer
      stage_id:
        type: string
    required:
    - character_id
    - music_id
    - score
    - stage_id
    type: object
  entity.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
    type: object
  entity.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
  entity.SessionResponse:
    properties:
      session_id:
        type: string
      user:
        $ref: '#/definitions/entity.UserResponse'
    type: object
  entity.StoveSignInRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
    required:
    - avatar_url
    - display_name
    - email
    - id
    type: object
  entity.UserResponse:
    properties:
      coin:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      exp:
        type: integer
      gem:
        type: integer
      id:
        type: string
      is_verified:
        type: boolean
      language:
        type: string
      level:
        type: integer
      nickname:
        type: string
      platform_avatar_url:
        type: string
      platform_display_name:
        type: string
      platform_email:
        type: string
      platform_type:
        type: string
      platform_user_id:
        type: string
      updated_at:
        type: string
    type: object
  repository.BestRecordResponse:
    properties:
      record:
        $ref: '#/definitions/repository.RecordResponse'
    type: object
  repository.ListRecordResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/repository.RecordResponse'
        type: array
    type: object
  repository.RecordResponse:
    properties:
      accuracy:
        type: number
      additional_info:
        additionalProperties: true
        type: object
      bad_count:
        type: integer
      character_id:
        type: string
      created_at:
        type: string
      good_count:
        type: integer
      id:
        type: string
      is_full_combo:
        type: boolean
      is_perfect_play:
        type: boolean
      max_combo:
        type: integer
      miss_count:
        type: integer
      music_id:
        type: string
      perfect_count:
        type: integer
      play_duration:
        type: integer
      played_at:
        type: string
      rank:
        type: string
      score:
        type: integer
      stage_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  repository.SingleRecordResponse:
    properties:
      record:
        $ref: '#/definitions/repository.RecordResponse'
    type: object
host: localhost:8080
info:
  contact:
    email: support@witchslounge.com
    name: API Support
  description: Witch's Lounge Backend API Server
  title: Witch's Lounge API
  version: "1.0"
paths:
  /api/v1/records:
    post:
      consumes:
      - application/json
      description: 플레이 기록을 저장합니다
      parameters:
      - description: 플레이 기록 요청
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreateRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 생성된 기록
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/repository.SingleRecordResponse'
              type: object
        "400":
          description: 잘못된 요청 형식
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: 인증 필요
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 플레이 기록 저장
      tags:
      - Record
  /musics:
    get:
      consumes:
      - application/json
      description: 활성화 된 악곡들을 가져옵니다
      produces:
      - application/json
      responses:
        "200":
          description: 성공
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MusicResponse'
                  type: array
              type: object
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 활성화 된 악곡들을 가져옵니다
      tags:
      - Musics
  /musics/{music_id}/stages:
    get:
      consumes:
      - application/json
      description: 음악 id에 따른 스테이지들을 가져옵니다
      parameters:
      - description: Music ID
        in: path
        name: music_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 성공
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StageResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "403":
          description: 음악이 비활성화됨
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 음악 id에 따른 스테이지들을 가져옵니다
      tags:
      - Stages
  /records:
    get:
      consumes:
      - application/json
      description: 해당 곡/스테이지의 전체 기록을 조회합니다
      parameters:
      - description: 음악 ID(UUID)
        in: query
        name: music_id
        required: true
        type: string
      - description: 스테이지 ID(UUID)
        in: query
        name: stage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 전체 기록
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/repository.ListRecordResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 전체 기록 조회
      tags:
      - Record
  /records/best:
    get:
      consumes:
      - application/json
      description: 해당 곡/스테이지의 최고 기록을 조회합니다
      parameters:
      - description: 음악 ID(UUID)
        in: query
        name: music_id
        required: true
        type: string
      - description: 스테이지 ID(UUID)
        in: query
        name: stage_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 최고 기록
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/repository.BestRecordResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 최고 기록 조회
      tags:
      - Record
  /stove/signin:
    post:
      consumes:
      - application/json
      description: Stove 플랫폼 사용자 정보로 로그인 처리 및 세션 생성
      parameters:
      - description: Stove 로그인 요청 정보
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.StoveSignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 로그인 성공 및 세션 정보
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.SessionResponse'
              type: object
        "400":
          description: 요청 데이터 검증 실패
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: Stove 플랫폼으로 로그인
      tags:
      - Stove
  /user/me:
    get:
      consumes:
      - application/json
      description: 세션 정보를 통해 현재 인증된 사용자의 상세 정보를 반환
      produces:
      - application/json
      responses:
        "200":
          description: 사용자 정보
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.UserResponse'
              type: object
        "401":
          description: 인증 정보 없음
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 현재 사용자 정보 조회
      tags:
      - User
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"